% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{aggregate_predictions}
\alias{aggregate_predictions}
\title{Title}
\usage{
aggregate_predictions(
  models_or_predictions,
  aggregation_method = "Median",
  data = NULL,
  ...
)
}
\arguments{
\item{models_or_predictions}{<- a vector which have elements of one of two forms, either a numeric length two vector containing the home score and away score of a prediction, or a function that takes in data (of the format explained in data) and returns a numeric length two vector containing the home score and away score}

\item{aggregation_method}{<- A string representing the name of the aggregation method (defaults to median)}

\item{data}{<- A list with the selcted statistics (the four factors for home and away, off and def efficency for home and away, and Pace/Tempo for home and away)}

\item{...}{<- additional arguments to be passed in our aggregator}
}
\value{
A numeric length two vector containing the home score and away score
}
\description{
Title
}
\examples{
#Game Data from TAMU 11/30/2024
game_data <- list(
eFG_pct_home = 47.2,
eFG_pct_away = 50.8,
TO_pct_home = 18.3,
TO_pct_away = 14.6,
OR_pct_home = 43.6,
OR_pct_away = 30.2,
FT_rate_home = .438,
FT_rate_away = .397,
offensive_efficiency_home = 114.3,
defensive_efficiency_home = 94.0,
offensive_efficiency_away = 111.9,
defensive_efficiency_away = 101.4,
pace_home = 67.0,
pace_away = 70.3
)
#Input model, three score predictions (VEGAS, KENPOM, and EvanMiya), and then our four functions
input = c(c(78, 72.5), c(76, 70), c(79, 74),
efficiency_model, log5_model, linear_reg_model, logistic_model)
aggregate_predictions(input, aggregation_method="Geometric Mean", data=game_data)
#Actual Score was 81-77

}
