% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregations.R
\name{bayesian_averaging}
\alias{bayesian_averaging}
\title{Bayesian Averaging}
\usage{
bayesian_averaging(prediction_matrix, priors)
}
\arguments{
\item{prediction_matrix}{Takes in a matrix with home scores in one column and away scores in the other}

\item{priors}{A numeric vector representing the prior probabilities for each prediction (row in the matrix).
The length of the priors vector must match the number of rows in the prediction matrix, and the priors must sum to 1.}
}
\value{
a numeric length two vector containing the home score and away score forecasts using bayesian averaging
}
\description{
Title
}
\examples{
# Example matrix of predictions
prediction_matrix <- matrix(c(78, 72, 76, 70, 79, 74), ncol = 2, byrow = TRUE)
priors <- c(0.3, 0.4, 0.3)

# Perform Bayesian averaging
bayesian_averaging(prediction_matrix, priors)
}
