% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregations.R
\name{geometric_mean}
\alias{geometric_mean}
\title{Geometric Mean}
\usage{
geometric_mean(prediction_matrix)
}
\arguments{
\item{prediction_matrix}{A matrix with home scores in one column and away scores in the other}
}
\value{
a numeric length two vector containg the aggregated home and away score forcasts using a geometric mean
}
\description{
Title
}
\examples{
prediction_matrix <- matrix(
  c(
    78, 72,
    80, 74,
    76, 70,
    82, 77,
    79, 75,
    81, 73,
    66, 58,
    77, 76,
    79, 74,
    80, 82

  ),
  ncol = 2,
  byrow = TRUE
)
trimmed_mean(prediction_matrix, trim = 0.1)
}
