% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{log5_model}
\alias{log5_model}
\title{Log5 Model}
\usage{
log5_model(data, pythagorean_param = 11.5)
}
\arguments{
\item{data}{A list with the selcted statistics (the four factors for home and away, off and def efficency for home and away, and Pace/Tempo for home and away)}

\item{pythagorean_param}{the exponenet in our Pythogorean Score function (highly reccommended to keep at default 11.5 for best behavior)}
}
\value{
numeric length two vector containing the home score and away score of the model
}
\description{
Title
}
\examples{
#Game Data from TAMU 11/30/2024
game_data <- list(
eFG_pct_home = 47.2,
eFG_pct_away = 50.8,
TO_pct_home = 18.3,
TO_pct_away = 14.6,
OR_pct_home = 43.6,
OR_pct_away = 30.2,
FT_rate_home = .438,
FT_rate_away = .397,
offensive_efficiency_home = 114.3,
defensive_efficiency_home = 94.0,
offensive_efficiency_away = 111.9,
defensive_efficiency_away = 101.4,
pace_home = 67.0,
pace_away = 70.3
)
log5_model(game_data)
#Actual Score was 81-77
}
