% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregations.R
\name{trimmed_mean}
\alias{trimmed_mean}
\title{Title}
\usage{
trimmed_mean(prediction_matrix, trim = 0.1)
}
\arguments{
\item{prediction_matrix}{Takes in a matrix with home scores in one column and away scores in the other}

\item{trim}{The proportion of scores you would like to remove before taking the mean}
}
\value{
a numeric length two vector containing the home score and away score forecasts using a trimmed mean
}
\description{
Title
}
\examples{
prediction_matrix <- matrix(
  c(
    78, 72,
    80, 74,
    76, 70,
    82, 77,
    79, 75,
    81, 73,
    66, 58,
    77, 76,
    79, 74,
    80, 82

  ),
  ncol = 2,
  byrow = TRUE
)
trimmed_mean(prediction_matrix, trim = 0.1)
}
