% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregations.R
\name{exponential_smoothing}
\alias{exponential_smoothing}
\title{Title}
\usage{
exponential_smoothing(prediction_matrix, alpha = 0.5)
}
\arguments{
\item{prediction_matrix}{Takes in a matrix with home scores in one column and away scores in the other}

\item{alpha}{}
}
\description{
Title
}
\examples{
prediction_matrix <- matrix(
  c(
    78, 72,
    80, 74,
    76, 70,
    82, 77,
    79, 75,
    81, 73,
    66, 58,
    77, 76,
    79, 74,
    80, 82

  ),
  ncol = 2,
  byrow = TRUE
)
exponential_smoothing(prediction_matrix)
}
