% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacleaningfunction.R
\name{wrangle_basketball_data}
\alias{wrangle_basketball_data}
\title{Wrangle Basketball Data}
\usage{
wrangle_basketball_data(basketball_data)
}
\arguments{
\item{basketball_data}{<- takes in a dataset of the form of that recieved from hoopR Data (https://github.com/sportsdataverse/hoopR-data/tree/main)}
}
\value{
Returns the dataset in a form that works for our aggregation and model functions, does so by computing a variety of metrics and cleaning up the data
}
\description{
Title
}
\examples{
library(dplyr)
library(tidyr)
#reading in our sample data
file_path <- system.file("extdata", "team_box_2022.csv", package = "NCAAggregator")
basketball_data <- read.csv(file_path)
wrangle_basketball_data(basketball_data)
}
